extract controller methods and associated tests to do with the delete logs feature into their own controller,     amend routes accordingly
